Extract and export the `string` serializer
