+++ packages/umi-serializer-beet/src/createBeetSerializer.ts
@@ -1,23 +1,14 @@
 import {
   BoolSerializerOptions,
-  fixSerializer,
-  mergeBytes,
   none,
   Serializer,
   SerializerInterface,
-  StringSerializerOptions,
   UnitSerializerOptions,
-  utf8,
 } from '@metaplex-foundation/umi';
 import { array } from './array';
 import { bytes } from './bytes';
 import { dataEnum } from './dataEnum';
-import {
-  BeetSerializerError,
-  DeserializingEmptyBufferError,
-  NotEnoughBytesError,
-} from './errors';
-import { getSizeDescription } from './getSizeDescription';
+import { BeetSerializerError, DeserializingEmptyBufferError } from './errors';
 import { map } from './map';
 import { nullable } from './nullable';
 import {
@@ -38,6 +29,7 @@ import { option } from './option';
 import { publicKey } from './pubkey';
 import { scalarEnum } from './scalarEnum';
 import { set } from './set';
+import { string } from './string';
 import { struct } from './struct';
 import { tuple } from './tuple';
 
@@ -73,50 +65,6 @@ function getTolerantSerializerFactory<
 export function createBeetSerializer(
   options: BeetSerializerOptions = {}
 ): SerializerInterface {
-  const string = (
-    options: StringSerializerOptions = {}
-  ): Serializer<string> => {
-    const size = options.size ?? u32();
-    const encoding = options.encoding ?? utf8;
-    const description =
-      options.description ??
-      `string(${encoding.description}; ${getSizeDescription(size)})`;
-
-    if (size === 'variable') {
-      return { ...encoding, description };
-    }
-
-    if (typeof size === 'number') {
-      return fixSerializer(encoding, size, description);
-    }
-
-    return {
-      description,
-      fixedSize: null,
-      maxSize: null,
-      serialize: (value: string) => {
-        const contentBytes = encoding.serialize(value);
-        const lengthBytes = size.serialize(contentBytes.length);
-        return mergeBytes([lengthBytes, contentBytes]);
-      },
-      deserialize: (buffer: Uint8Array, offset = 0) => {
-        if (buffer.slice(offset).length === 0) {
-          throw new DeserializingEmptyBufferError('string');
-        }
-        const [lengthBigInt, lengthOffset] = size.deserialize(buffer, offset);
-        const length = Number(lengthBigInt);
-        offset = lengthOffset;
-        const contentBuffer = buffer.slice(offset, offset + length);
-        if (contentBuffer.length < length) {
-          throw new NotEnoughBytesError('string', length, contentBuffer.length);
-        }
-        const [value, contentOffset] = encoding.deserialize(contentBuffer);
-        offset += contentOffset;
-        return [value, offset];
-      },
-    };
-  };
-
   const bool = (options: BoolSerializerOptions = {}): Serializer<boolean> => {
     const size = options.size ?? u8();
     if (size.fixedSize === null) {
