+++ packages/umi-serializer-beet/src/index.ts
@@ -8,6 +8,7 @@ export * from './nullable';
 export * from './option';
 export * from './scalarEnum';
 export * from './set';
+export * from './string';
 export * from './struct';
 export * from './tuple';
 export * from './plugin';
