+++ packages/umi-serializer-beet/test/nullable.test.ts
@@ -2,9 +2,10 @@ import test from 'ava';
 import { nullable } from '../src/nullable';
 import { createBeetSerializer } from '../src';
 import { d, s } from './_helpers';
+import { string } from '../src/string';
 
 test('regular (de)serialization', (t) => {
-  const { u8, u16, u64, string } = createBeetSerializer();
+  const { u8, u16, u64 } = createBeetSerializer();
 
   // Null.
   s(t, nullable(u8()), null, '00');
@@ -35,7 +36,7 @@ test('regular (de)serialization', (t) => {
 });
 
 test('fixed (de)serialization', (t) => {
-  const { u8, u16, u64, string } = createBeetSerializer();
+  const { u8, u16, u64 } = createBeetSerializer();
   const fixedU8 = nullable(u8(), { fixed: true });
   const fixedU8WithU16Prefix = nullable(u8(), { fixed: true, prefix: u16() });
   const fixedString = nullable(string({ size: 5 }), { fixed: true });
@@ -77,7 +78,7 @@ test('fixed (de)serialization', (t) => {
 });
 
 test('description', (t) => {
-  const { u8, u16, string } = createBeetSerializer();
+  const { u8, u16 } = createBeetSerializer();
   t.is(nullable(u8()).description, 'nullable(u8; u8)');
   t.is(nullable(string()).description, 'nullable(string(utf8; u32(le)); u8)');
   t.is(nullable(u8(), { prefix: u16() }).description, 'nullable(u8; u16(le))');
@@ -101,7 +102,7 @@ test('description', (t) => {
 });
 
 test('sizes', (t) => {
-  const { u8, u16, string } = createBeetSerializer();
+  const { u8, u16 } = createBeetSerializer();
   t.is(nullable(u8()).fixedSize, null);
   t.is(nullable(u8()).maxSize, 2);
   t.is(nullable(string()).fixedSize, null);
