+++ packages/umi-serializer-beet/test/struct.test.ts
@@ -3,9 +3,10 @@ import { option } from '../src/option';
 import { createBeetSerializer } from '../src';
 import { s, d } from './_helpers';
 import { struct } from '../src/struct';
+import { string } from '../src/string';
 
 test('(de)serialization', (t) => {
-  const { u8, u64, string } = createBeetSerializer();
+  const { u8, u64 } = createBeetSerializer();
 
   // Empty struct.
   s(t, struct([]), {}, '');
@@ -33,7 +34,7 @@ test('(de)serialization', (t) => {
 });
 
 test('description', (t) => {
-  const { u8, string } = createBeetSerializer();
+  const { u8 } = createBeetSerializer();
   t.is(struct([['age', u8()]]).description, 'struct(age: u8)');
   t.is(
     struct([
@@ -49,7 +50,7 @@ test('description', (t) => {
 });
 
 test('sizes', (t) => {
-  const { u8, u64, string } = createBeetSerializer();
+  const { u8, u64 } = createBeetSerializer();
   t.is(struct([]).fixedSize, 0);
   t.is(struct([]).maxSize, 0);
   t.is(struct([['age', u8()]]).fixedSize, 1);
