+++ packages/umi-serializer-data-view/test/array.test.ts
@@ -3,9 +3,10 @@ import { Endian } from '@metaplex-foundation/umi';
 import { array } from '../src/array';
 import { createDataViewSerializer } from '../src';
 import { s, d } from './_helpers';
+import { string } from '../src/string';
 
 test('prefixed (de)serialization', (t) => {
-  const { u8, string, u64 } = createDataViewSerializer();
+  const { u8, u64 } = createDataViewSerializer();
 
   // Empty.
   s(t, array(u8()), [], '00000000'); // 4-bytes prefix.
@@ -31,7 +32,7 @@ test('prefixed (de)serialization', (t) => {
 });
 
 test('fixed (de)serialization', (t) => {
-  const { u8, string, u64 } = createDataViewSerializer();
+  const { u8, u64 } = createDataViewSerializer();
 
   // Empty.
   s(t, array(u8(), { size: 0 }), [], '');
@@ -63,7 +64,7 @@ test('fixed (de)serialization', (t) => {
 });
 
 test('remainder (de)serialization', (t) => {
-  const { u8, string, u64 } = createDataViewSerializer();
+  const { u8, u64 } = createDataViewSerializer();
   const remainder = { size: 'remainder' } as const;
 
   // Empty.
@@ -112,7 +113,7 @@ test('description', (t) => {
 });
 
 test('sizes', (t) => {
-  const { u8, u32, string } = createDataViewSerializer();
+  const { u8, u32 } = createDataViewSerializer();
   t.is(array(u8()).fixedSize, null);
   t.is(array(u8()).maxSize, null);
   t.is(array(u8(), { size: u8() }).fixedSize, null);
