+++ packages/umi-serializer-data-view/test/struct.test.ts
@@ -2,10 +2,11 @@ import test from 'ava';
 import { option } from '../src/option';
 import { createDataViewSerializer } from '../src';
 import { s, d } from './_helpers';
+import { string } from '../src/string';
 import { struct } from '../src/struct';
 
 test('(de)serialization', (t) => {
-  const { u8, u64, string } = createDataViewSerializer();
+  const { u8, u64 } = createDataViewSerializer();
 
   // Empty struct.
   s(t, struct([]), {}, '');
@@ -33,7 +34,7 @@ test('(de)serialization', (t) => {
 });
 
 test('description', (t) => {
-  const { u8, string } = createDataViewSerializer();
+  const { u8 } = createDataViewSerializer();
   t.is(struct([['age', u8()]]).description, 'struct(age: u8)');
   t.is(
     struct([
@@ -49,7 +50,7 @@ test('description', (t) => {
 });
 
 test('sizes', (t) => {
-  const { u8, u64, string } = createDataViewSerializer();
+  const { u8, u64 } = createDataViewSerializer();
   t.is(struct([]).fixedSize, 0);
   t.is(struct([]).maxSize, 0);
   t.is(struct([['age', u8()]]).fixedSize, 1);
